+++ CHANGELOG.md
@@ -2692,3 +2692,4 @@
 - [2026-07-11 12:12] feat(navigation): integrate custom telemetry logger in navigation
 - [2026-07-11 12:25] fix(auth): prevent double tap duplicate action in auth
 - [2026-07-11 12:46] fix(core): fix memory leak by cancelling core subscription
+- [2026-07-11 12:50] style(cache): harmonize light and dark theme palette for cache